requirements.txt
click==8.1.7
dlib==19.24.2
emoji==2.9.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==9.4.0
pilmoji==2.0.4
